requirements.txt
Flask
langchain
faiss-cpu
python-dotenv
PyPDF2
google-generativeai
streamlit
SpeechRecognition
pyaudio
langchain-community
langchain-core 
langchain-google-genai
pypdf 
markdown